refac: add setup server and db before run test
