+++ .mocharc.json
@@ -1,5 +1,6 @@
 {
-  "spec": "test/**/*.test.ts",
+  "spec": "test/index.ts",
   "timeout": 5000,
-  "recursive": true
+  "recursive": true,
+  "node-option": ["import=tsx"]
 }
