+++ package.json
@@ -15,7 +15,8 @@
   "dependencies": {
     "@prisma/client": "6.6.0",
     "bcrypt": "5.1.1",
-    "fastify": "5.2.2"
+    "fastify": "5.2.2",
+    "dotenv": "16.5.0"
   },
   "type": "module",
   "devDependencies": {
@@ -26,7 +27,7 @@
     "eslint": "9.24.0",
     "eslint-config-prettier": "10.1.1",
     "eslint-plugin-prettier": "5.2.6",
-    "mocha": "^11.1.0",
+    "mocha": "11.1.0",
     "prettier": "3.5.3",
     "prisma": "6.6.0",
     "tsx": "4.19.3",
