--- src/shared/helper/user-validation.ts
@@ -1,5 +1,0 @@
-export function isStrongPassword(password: string) {
-    const hasLetter = /[a-zA-Z]/.test(password);
-    const hasDigit = /\d/.test(password);
-    return password.length >= 6 && hasLetter && hasDigit;
-}
