--- test/index.test.ts
@@ -1,10 +1,0 @@
-import assert from "assert";
-import { describe } from "node:test";
-
-describe('Array', function () {
-  describe('#indexOf()', function () {
-    it('should return -1 when the value is not present', function () {
-      assert.equal([1, 2, 3].indexOf(4), -1);
-    });
-  });
-});
